requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1


torch>=2.3.0  # Ensures a compatible PyTorch version

# Document processing
docling==1.3.0
pytesseract==0.3.10
Pillow>=10.2.0
PyPDF2==3.0.1
pdf2image==1.17.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File storage
boto3==1.34.0
minio==7.2.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
loguru==0.7.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
